[Rebuild] id의 type을 Int로 수정

+++ travelPlan/Source/Data/Network/Server/RequestDTO/DestinationScrapUpdateRequestDTO.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct DestinationScrapUpdateRequestDTO: Encodable {
-  let objectIdList: [Int64]
+  let objectIdList: [Int]
   let folderName: String
 }
